continued work on database transition - debugging still needed

+++ lib/UI/components/set_widget.dart
@@ -30,7 +30,7 @@ class SetWidget extends StatelessWidget {
               ],
               keyboardType: TextInputType.number,
               onChanged: (changes) {
-                thisSet.selectWeight = int.parse(changes);
+                thisSet.selectWeight = double.parse(changes);
               },
             ),
           ),
